pyproject.toml
[project]
name = "llm-fight-club"
version = "0.1.0"
description = "Multi-LLM group chat system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=0.1.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_fight_club"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
